fix: refactored checkboxes
