+++ src/lib/list/Info.svelte
@@ -1,8 +1,8 @@
 <script>
     import {Card, Accordion, AccordionItem} from 'flowbite-svelte'
-    import InfoView from './InfoView.svelte';
-    import ContractView from './ContractView.svelte';
-    import {Selected, Expanded} from "./selected";
+    import InfoView from '../InfoView.svelte';
+    import ContractView from '../ContractView.svelte';
+    import {Selected, SelectedExpanded} from "../selected";
 
 
     //We want the accordion to be open by default if there is only one item
@@ -16,7 +16,7 @@
     <Accordion multiple class="w-60">
 
         {#each $Selected as item, i}
-          <AccordionItem bind:open={$Expanded[i]}>
+          <AccordionItem bind:open={$SelectedExpanded[i]}>
             <div slot="header">{item.name}</div>
             <InfoView {item} />
             <br>
