+++ src/lib/selected.ts
@@ -9,10 +9,15 @@ interface Item {
 
 type ItemType = "Client" | "Location" | "Device" | "Call" | "Contract";
 
+//All selected elements
 const Selected:Writable<Item[]> = writable([]);
 
-const Expanded:Writable<boolean[]> = writable([]);
+//All selected elements as an array of whether they are expanded or not
+const SelectedExpanded:Writable<boolean[]> = writable([]);
+
+//All selected elements as an array of their backendIDs
+const SelectedID:Writable<string[]> = writable([]);
 
 
-export  { Selected, Expanded};
+export  { Selected, SelectedExpanded, SelectedID};
 export type { Item, ItemType };
